+++ packages/cohere/src/convert-to-cohere-chat-prompt.ts
@@ -47,10 +47,6 @@ export function convertToCohereChatPrompt(
               text += part.text;
               break;
             }
-            case 'redacted-reasoning':
-            case 'reasoning': {
-              break; // ignored
-            }
             case 'tool-call': {
               toolCalls.push({
                 id: part.toolCallId,
@@ -61,10 +57,6 @@ export function convertToCohereChatPrompt(
                 },
               });
               break;
-            }
-            default: {
-              const _exhaustiveCheck: never = part;
-              throw new Error(`Unsupported part: ${_exhaustiveCheck}`);
             }
           }
         }