+++ packages/cerebras/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/cerebras
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - @ai-sdk/provider@1.0.12
+  - @ai-sdk/openai-compatible@0.1.16
+  - @ai-sdk/provider-utils@2.1.14
+
 ## 0.1.15
 
 ### Patch Changes